benchmark: ValueError if invalid parameters are specified.
