+++ domainlab/arg_parser.py
@@ -3,12 +3,13 @@ Command line arguments
 """
 import argparse
 import warnings
+
 import yaml
 
 from domainlab.algos.compos.matchdg_args import add_args2parser_matchdg
 from domainlab.algos.trainers.args_dial import add_args2parser_dial
-from domainlab.models.args_vae import add_args2parser_vae
 from domainlab.models.args_jigen import add_args2parser_jigen
+from domainlab.models.args_vae import add_args2parser_vae
 
 
 def mk_parser_main():
@@ -158,6 +159,33 @@ def mk_parser_main():
     return parser
 
 
+def apply_dict_to_args(args, data: dict, extend=False):
+    """
+    Tries to apply the data to the args dict of DomainLab.
+    Unknown keys are silently ignored as long as
+    extend is not set.
+    # FIXME: do we have a test to ensure args dict from
+    # domainlab really got what is passed from "data" dict?
+    """
+    arg_dict = args.__dict__
+    for key, value in data.items():
+        if (key in arg_dict) or extend:
+            if isinstance(value, list):
+                cur_val = arg_dict.get(key, None)
+                if not isinstance(cur_val, list):
+                    if cur_val is not None:
+                        # @FIXME: should we warn or raise Error?
+                        warnings.warn(f"input dictionary value is list, \
+                                    however, in DomainLab args, we have {cur_val}, \
+                                    going to overrite to list")
+                    arg_dict[key] = []
+                arg_dict[key].extend(value)
+            else:
+                arg_dict[key] = value
+        else:
+            raise ValueError("Unsupported key: ", key)
+
+
 def parse_cmd_args():
     """
     get args from command line
@@ -165,20 +193,9 @@ def parse_cmd_args():
     parser = mk_parser_main()
     args = parser.parse_args()
     if args.config_file:
-
         data = yaml.safe_load(args.config_file)
         delattr(args, 'config_file')
-        arg_dict = args.__dict__
-
-        for key in data:
-            if key not in arg_dict:
-                raise ValueError("The key is not supported: ", key)
-
-        for key, value in data.items():
-            if isinstance(value, list):
-                arg_dict[key].extend(value)
-            else:
-                arg_dict[key] = value
+        apply_dict_to_args(args, data)
 
     if args.acon is None:
         print("\n\n")
