+++ tests/test_argparse.py
@@ -2,10 +2,12 @@
 Test argparser functionality
 """
 
-import sys
 import os
+import sys
+
 import pytest
-from domainlab.arg_parser import parse_cmd_args
+
+from domainlab.arg_parser import parse_cmd_args, mk_parser_main, apply_dict_to_args
 
 
 def test_parse_cmd_args_warning():
@@ -44,3 +46,13 @@ def test_parse_invalid_yml_args():
     
     with pytest.raises(ValueError):
         parse_cmd_args()
+
+
+def test_apply_dict_to_args():
+    """Testing apply_dict_to_args"""
+    parser = mk_parser_main()
+    args = parser.parse_args(args=[])
+    data = {'a': 1, 'b': [1, 2], 'aname': 'diva'}
+    apply_dict_to_args(args, data, extend=True)
+    assert args.a == 1
+    assert args.aname == 'diva'
